on 32 bits, drop high 32 bits of siphash

+++ src/libponyrt/ds/fun.c
@@ -67,12 +67,12 @@ static uint64_t siphash24(const unsigned char* key, const char* in, size_t len)
 
 size_t hash_block(const void* p, size_t len)
 {
-  return siphash24(the_key, (const char*)p, len);
+  return (size_t)siphash24(the_key, (const char*)p, len);
 }
 
 size_t hash_str(const char* str)
 {
-  return siphash24(the_key, str, strlen(str));
+  return (size_t)siphash24(the_key, str, strlen(str));
 }
 
 size_t hash_ptr(const void* p)
